initial settings enable handled
